Suspend Preview functionality added
Added the ability to suspend the live preview feed via a toggle button.
This is useful if images need to be captured but the user doesn't want
to deal with resources needed for managing the live stream.   Also
useful for systems where the live stream can cause freezing or crashes
if running for too long.
Also fixed a small bug in the window resize function.